+++ app/src/main/java/woowacourse/movie/domain/model/Movie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.domain.model
 
 data class Movie(
     val posterImageId: Int,
