+++ app/src/main/java/woowacourse/movie/domain/model/MovieTicket.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.domain.model
 
 class MovieTicket(
     val movieTitle: String,
