+++ app/src/main/java/woowacourse/movie/presentation/presenter/MainPresenterImpl.kt
@@ -1,6 +1,8 @@
-package woowacourse.movie
+package woowacourse.movie.presentation.presenter
 
-import woowacourse.movie.model.Movie
+import woowacourse.movie.R
+import woowacourse.movie.domain.model.Movie
+import woowacourse.movie.presentation.contract.MainContract
 
 class MainPresenterImpl(private val view: MainContract.View) : MainContract.Presenter {
     override fun createMovieList(): List<Movie> {
