+++ app/src/main/java/woowacourse/movie/presentation/view/MainActivity.kt
@@ -1,8 +1,13 @@
-package woowacourse.movie
+package woowacourse.movie.presentation.view
 
 import android.content.Intent
 import android.widget.ListView
-import woowacourse.movie.model.Movie
+import woowacourse.movie.presentation.contract.MainContract
+import woowacourse.movie.presentation.presenter.MainPresenterImpl
+import woowacourse.movie.presentation.adapter.MovieListAdapter
+import woowacourse.movie.R
+import woowacourse.movie.domain.model.Movie
+import woowacourse.movie.presentation.base.BaseActivity
 
 class MainActivity : BaseActivity(), MainContract.View {
     private lateinit var adapter: MovieListAdapter
