+++ app/src/test/java/woowacourse/movie/model/MovieTest.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import woowacourse.movie.domain.model.Movie
 
 class MovieTest {
     @Test
